requirements.txt
openai-whisper
torch
fastapi
uvicorn[standard]
python-multipart
ffmpeg-python
gradio
tqdm
aiofiles